match: ~67 functions byte-matched across 6 regions (pattern-DB-triaged fan-out)
6 parallel matching agents, each using pattern_suggest (triage + oracle-shape
isolation) plus the methodology lever catalog and raw-oracle reading. All
pin-free, every seal verify_asm=PASS, full-recon coherence gate clean (no
cross-module breakage).

- eaclib/sndpsxz (14): SPU-register family via volatile struct-base reload;
  magic-divide RE (SNDfxlevel /16129 = /127^2); (signed char) guard reads.
- eaclib/eacpsxz (10): timer.cpp + xform.cpp whole TUs; gp-rel ownership
  (ticks absolute vs gTickVal gp-rel); index-vs-pointer hoist; redundant-andi drop.
- game/common spchevnt (26): the whole SPCHNFS_* family is now 53/53 --
  ascending parms[] assignment order fills the load-delay slot (one lever, all 26).
- game/common+psx (14): vtable 8-byte-stride decay (char* + raw offset);
  Cars_gList **->*[] unsized-array; branch-polarity; plus a CarIO link-bug fix.
- frontend (2 sealed + 2 reduced 7->1): goto-done cross-jump-merge;
  char-capture idiom from a sealed sibling (pattern_suggest @128.7).
- syslib (CD_initvol): nested-if short-circuit + assignment-chain order flip.

pattern_suggest verdict (the tool under test): its divergence-shape isolation
and score-as-triage were consistently useful (high score = a sealed-sibling
lead to read, low/null = a floor to skip); its sealed-sibling SOURCE rarely
drove the fix directly, because a high score usually lands on the shared
already-matching shape. It accelerates triage; it does not replace the lever
catalog + raw oracle.

+++ recon/game/common/AIWORLD.cpp
@@ -280,7 +280,7 @@ int AIWorld_GetProfileMask(int laneIndex)
 /* ---- AIWorld_IsDriveableLane_UsingMask__Fii  [@0x8007343c] ---- */
 int AIWorld_IsDriveableLane_UsingMask(int slice,int mask)
 {
-  return *(short *)(slice * 0x20 + BWorldSm_slices + 0x16) & mask;
+  return *(short *)(slice * 0x20 + (int)BWorldSm_slices + 0x16) & mask;
 }
 
 /* ---- AIWorld_CheckForBarrierBetweenLanes__Fiii  [@0x80073458] ---- */
